requirements.txt
llvmlite==0.39.0
networkx==2.8.6
numba==0.56.0
numpy==1.22.4
pip==21.1.3
python-dotenv==0.20.0
setuptools==56.0.0
